backend/requirements.txt
asgiref==3.8.1
attrs==25.1.0
colorama==0.4.6
dj-database-url==2.3.0
Django==4.2
django-cors-headers==4.4.0
django-crum==0.7.9
django-jazzmin==3.0.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-spectacular-sidecar==2024.12.1
et_xmlfile==2.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
numpy==1.26.3
openpyxl==3.1.5
pandas==2.2.3
pillow==11.1.0
psycopg==3.2.4
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
qrcode==8.0
referencing==0.36.2
rpds-py==0.22.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
